pyproject.toml
[project]
name = "unetr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch",
    "torchvision",
    "pyyaml>=6.0.2",
    "tifffile>=2024.8.30",
    "monai>=1.4.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.12.0",
    "einops>=0.8.0",
]


[tool.ruff]
extend-select = ["I"]
exclude = [".venv"]

[tool.uv.sources]
torch = [{ index = "pytorch-cu126" }]
torchvision = [{ index = "pytorch-cu126" }]


[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
